refactor: introduce provider pattern to alert code
